chore(telescope): separete 'recents' into its own module
